+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed return type of `getDefaultCollectionFields` and custom collection `fields`
 - Renamed `typesenseOptions` to `serverConfig` in PluginOptions and add doc comments.
+- Added VN locale.
 
 ## 0.0.1